requirements.txt
dash
dash-bootstrap-components
plotly
numpy
opencv-python
pillow
pyyaml
pandas
ultralytics
torch
torchvision
